school-project/backend/requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.1
